requirement.txt
streamlit
pandas
google.generativeai
python-dotenv